move a bunch of code from the cli to their own files.
